ref: frontend build setup to copy files from s3 bucket – AH-2648 (#6)
Previously we were building the frontend web component of the app in
Heroku which took about 8mins. Now we build the frontend in CI and copy
over the static files to the Heroku host and then unpack and copy the files
we want to the public s3 bucket.

This change adds an additional ENV var to the script:

`S3_AH_BUILD_ARTIFACTS_BUCKET_NAME`

which is our private bucket for build artifacts.


[AH-2648]

[AH-2648]: https://admithubteam.atlassian.net/browse/AH-2648